requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
Flask==1.1.1
Flask-Session==0.3.1
Flask-SocketIO==4.2.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pipenv==2018.11.26
psycopg2-binary==2.8.3
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.7
sqlparse==0.3.0
urllib3==1.25.8
virtualenv==16.7.9
virtualenv-clone==0.5.3
Werkzeug==0.15.5
uuid